requirements.txt
Flask==0.10.1
gunicorn==19.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pymongo==3.2.2
python-dateutil==2.5.2
requests==2.7.0
six==1.10.0
Werkzeug==0.11.5
wheel==0.24.0
xmltodict==0.10.1
paho-mqtt==1.1
